+++ lib/public/components/NavBar/index.js
@@ -17,7 +17,7 @@ import { iconPerson, iconPlus } from '/js/src/icons.js';
 import { userPreferencesStore } from '../../utilities/userPreferencesStore.js';
 import { frontLink } from '../common/navigation/frontLink.js';
 import { absoluteFrontLink } from '../common/navigation/absoluteFrontLink.js';
-import { externalLinks } from '../common/externalLinks.js';
+import { externalLinks } from '../common/navigation/externalLinks.js';
 
 let isChecked = true;
