+++ lib/public/components/common/navigation/externalLinks.js
